+++ app/domain/cities/__init__.py
@@ -1,0 +1,10 @@
+from .entities import City
+from .errors import CityAlreadyExistsError, CityError, CityNotFoundError, CityValidationError
+
+__all__ = [
+    "City",
+    "CityAlreadyExistsError",
+    "CityError",
+    "CityNotFoundError",
+    "CityValidationError",
+]