+++ app/routes/__init__.py
@@ -5,7 +5,6 @@ from app.web import register_namespaces as register_web_namespaces
 
 from .account_route import account_bp
 from .namespaces.api_key_ns import api_key_ns
-from .namespaces.city_ns import city_ns
 from .namespaces.login_ns import login_ns
 from .namespaces.object_ns import object_ns
 from .namespaces.object_status_ns import object_status_ns
@@ -23,7 +22,6 @@ def register_routes(app: Flask):
 def register_namespaces(api: Api):
     api.add_namespace(login_ns)
     api.add_namespace(api_key_ns)
-    api.add_namespace(city_ns)
     api.add_namespace(object_status_ns)
     api.add_namespace(object_ns)
     api.add_namespace(project_ns)